+++ entities/src/units/buildings/towers/tower.h
@@ -12,7 +12,7 @@
 class Tower : public Attackable {
    protected:
     unsigned int time_of_last_attack_;
-    int level_;
+    size_t level_;
     size_t max_level_;
     std::vector<unsigned int> max_health_per_level_;
     std::vector<unsigned int> attack_cooldown_per_level_;
@@ -29,7 +29,7 @@ class Tower : public Attackable {
           std::vector<size_t> &repair_cost_per_level,
           std::vector<size_t> &upgrade_cost_per_level,
           std::vector<double> &attack_radius_per_level,
-          unsigned int current_time, Coordinate position, int level);
+          unsigned int current_time, Coordinate position, size_t level);
     virtual void attack(std::vector<Attackable> &enemies) = 0;
     virtual bool isReadyForAttack(unsigned int current_time);
     virtual std::vector<Attackable> findTargets(
